pyproject.toml
[tool.pyink]
line-length = 80
preview = true
pyink-indentation = 2
pyink-use-majority-quotes = true

[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
    "google-auth==2.3.3",
    "google-api-python-client>=2.111.0",
    "oauth2client==4.1.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=0.5.3",
]
requires-python = ">=3.10"
